+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   chunk if they arent allowed. This is workaround for hytale's interaction bugs #40 #19
 * Added extra precautions for when file loading fails and added backups of those files
 * Added a chunk border system that will show particles when close to an edge of a claimed chunk, closes #79
+* Added a config option to configure blocks that will ignore all the restrictions
 
 # 1.0.14